begin refactoring ui folder
