+++ lib/ui/core/permission_denied_dialog.dart
