+++ lib/ui/create_medication_view.dart
@@ -4,7 +4,7 @@ import 'package:medication_tracker/data/model/medication_model.dart';
 import 'package:medication_tracker/data/providers/medication_provider.dart';
 import 'package:medication_tracker/data/providers/profile_provider.dart';
 import 'package:medication_tracker/ui/core/black_button.dart';
-import 'package:medication_tracker/widgets/header.dart';
+import 'package:medication_tracker/ui/core/header.dart';
 import 'package:provider/provider.dart';
 
 class CreateMedicationPage extends StatefulWidget {
