+++ lib/ui/edit_medication/full_screen_image_view.dart
