+++ lib/ui/edit_medication/zoomable_image.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 import 'package:flutter/material.dart';
-import 'package:medication_tracker/ui/full_screen_image_view.dart'; // Make sure this path is correct
+import 'package:medication_tracker/ui/edit_medication/full_screen_image_view.dart'; // Make sure this path is correct
 
 class ZoomableImage extends StatelessWidget {
   final String imagePath;
