+++ lib/ui/edit_profile_view.dart
@@ -4,8 +4,8 @@ import 'package:mask_text_input_formatter/mask_text_input_formatter.dart';
 import 'package:medication_tracker/data/model/user_profile_model.dart';
 import 'package:medication_tracker/data/providers/profile_provider.dart';
 import 'package:medication_tracker/ui/core/black_button.dart';
-import 'package:medication_tracker/widgets/header.dart';
-import 'package:medication_tracker/widgets/privacy_policy_button.dart';
+import 'package:medication_tracker/ui/core/header.dart';
+import 'package:medication_tracker/ui/core/privacy_policy_button.dart';
 import 'package:provider/provider.dart';
 
 class EditProfilePage extends StatefulWidget {
