+++ lib/ui/home/med_tile.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:medication_tracker/data/model/medication_model.dart';
 import 'package:medication_tracker/data/providers/medication_provider.dart';
-import 'package:medication_tracker/ui/edit_medication_view.dart';
+import 'package:medication_tracker/ui/edit_medication/edit_medication_view.dart';
 import 'package:provider/provider.dart';
 
 class MedicationTile extends StatelessWidget {
