+++ lib/ui/home/nav_bar.dart
