+++ lib/ui/select_profile_view.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:medication_tracker/data/providers/profile_provider.dart';
 import 'package:medication_tracker/ui/create_profile_view.dart';
-import 'package:medication_tracker/ui/home_view.dart';
+import 'package:medication_tracker/ui/home/home_view.dart';
 import 'package:medication_tracker/ui/core/black_button.dart';
-import 'package:medication_tracker/widgets/header.dart';
+import 'package:medication_tracker/ui/core/header.dart';
 import 'package:provider/provider.dart';
 
 class SelectProfilePage extends StatelessWidget {
